chore: refactor ingredients to module
