+++ backend/src/recipes/dto.ts
@@ -1,10 +1,8 @@
 import {
-  CreateIngredientDto as sharedCreateIngredientDto,
   CreateRecipeDto as sharedCreateRecipeDto,
   UpdateRecipeDto as sharedUpdateRecipeDto,
 } from '@app/shared';
-import { Type } from 'class-transformer';
-import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
+import { IsNotEmpty, IsString } from 'class-validator';
 //
 export class CreateRecipeDto implements sharedCreateRecipeDto {
   @IsString()
@@ -29,18 +27,4 @@ export class UpdateRecipeDto implements sharedUpdateRecipeDto {
 
   @IsString()
   author!: string;
-}
-
-export class CreateIngredientDto implements sharedCreateIngredientDto {
-  @IsString()
-  @IsNotEmpty()
-  name!: string;
-
-  @Type(() => Number)
-  @IsNumber()
-  amount!: number;
-
-  @IsString()
-  @IsNotEmpty()
-  unit!: string;
 }
