+++ backend/src/recipes/recipe.entity.ts
@@ -6,7 +6,7 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import { IngredientEntity } from './ingredient.entity';
+import { IngredientEntity } from '../ingredients/ingredient.entity';
 
 @Entity('recipes')
 export class RecipeEntity {
