+++ backend/src/recipes/recipe.service.spec.ts
@@ -2,14 +2,12 @@ import { NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { IngredientEntity } from './ingredient.entity';
 import { RecipeEntity } from './recipe.entity';
 import { RecipeService } from './recipe.service';
 
 describe('RecipeService', () => {
   let service: RecipeService;
   let recipeRepo: jest.Mocked<Repository<RecipeEntity>>;
-  let ingredientRepo: jest.Mocked<Repository<IngredientEntity>>;
 
   const mockRecipe: RecipeEntity = {
     id: 1,
@@ -19,17 +17,6 @@ describe('RecipeService', () => {
     createdAt: new Date('2026-01-01'),
     updatedAt: new Date('2026-01-01'),
     ingredients: [],
-  };
-
-  const mockIngredient: IngredientEntity = {
-    id: 1,
-    recipeId: 1,
-    recipe: {} as RecipeEntity,
-    name: 'Milk',
-    amount: 0.5,
-    unit: 'Liters',
-    createdAt: new Date('2026-01-01'),
-    updatedAt: new Date('2026-01-01'),
   };
 
   beforeEach(async () => {
@@ -46,20 +33,11 @@ describe('RecipeService', () => {
             remove: jest.fn().mockResolvedValue(undefined),
           },
         },
-        {
-          provide: getRepositoryToken(IngredientEntity),
-          useValue: {
-            find: jest.fn().mockResolvedValue([mockIngredient]),
-            create: jest.fn().mockReturnValue([mockIngredient]),
-            save: jest.fn().mockResolvedValue([mockIngredient]),
-          },
-        },
       ],
     }).compile();
 
     service = module.get(RecipeService);
     recipeRepo = module.get(getRepositoryToken(RecipeEntity));
-    ingredientRepo = module.get(getRepositoryToken(IngredientEntity));
   });
 
   it('should be defined', () => {
@@ -117,44 +95,6 @@ describe('RecipeService', () => {
     it('should remove the todo', async () => {
       await service.remove(1);
       expect(recipeRepo.remove).toHaveBeenCalledWith(mockRecipe);
-    });
-  });
-
-  describe('getIngredients', () => {
-    it('should return ingredients for an existing recipe', async () => {
-      const result = await service.getIngredients(1);
-      expect(result).toEqual([mockIngredient]);
-      expect(ingredientRepo.find).toHaveBeenCalledWith({
-        where: { recipeId: 1 },
-        order: { id: 'ASC' },
-      });
-    });
-
-    it('should throw NotFoundException if recipe does not exist', async () => {
-      recipeRepo.findOneBy.mockResolvedValue(null);
-      await expect(service.getIngredients(99)).rejects.toThrow(NotFoundException);
-    });
-  });
-
-  describe('addIngredients', () => {
-    it('should add new ingredients and return the full list', async () => {
-      const dtoList = [
-        { name: 'Milk', amount: 0.5, unit: 'Liters' },
-        { name: 'Eggs', amount: 1, unit: 'pieces' },
-      ];
-
-      const result = await service.addIngredients(1, dtoList);
-
-      expect(ingredientRepo.create).toHaveBeenCalledWith([
-        { recipeId: 1, name: 'Milk', amount: 0.5, unit: 'Liters' },
-        { recipeId: 1, name: 'Eggs', amount: 1, unit: 'pieces' },
-      ]);
-      expect(ingredientRepo.save).toHaveBeenCalled();
-      expect(ingredientRepo.find).toHaveBeenCalledWith({
-        where: { recipeId: 1 },
-        order: { id: 'ASC' },
-      });
-      expect(result).toEqual([mockIngredient]);
     });
   });
 });
