+++ backend/src/recipes/recipe.service.ts
@@ -1,8 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { CreateIngredientDto, CreateRecipeDto, UpdateRecipeDto } from './dto';
-import { IngredientEntity } from './ingredient.entity';
+import { CreateRecipeDto, UpdateRecipeDto } from './dto';
 import { RecipeEntity } from './recipe.entity';
 
 @Injectable()
@@ -10,8 +9,6 @@ export class RecipeService {
   constructor(
     @InjectRepository(RecipeEntity)
     private readonly recipeRepo: Repository<RecipeEntity>,
-    @InjectRepository(IngredientEntity)
-    private readonly ingredientRepo: Repository<IngredientEntity>,
   ) {}
 
   async findAll(): Promise<RecipeEntity[]> {
@@ -42,29 +39,5 @@ export class RecipeService {
   async remove(id: number): Promise<void> {
     const recipe = await this.findOne(id);
     await this.recipeRepo.remove(recipe);
-  }
-
-  async getIngredients(recipeId: number): Promise<IngredientEntity[]> {
-    await this.findOne(recipeId);
-    return this.ingredientRepo.find({
-      where: { recipeId },
-      order: { id: 'ASC' },
-    });
-  }
-
-  async addIngredients(recipeId: number, dtos: CreateIngredientDto[]): Promise<IngredientEntity[]> {
-    await this.findOne(recipeId);
-
-    const ingredients = this.ingredientRepo.create(
-      dtos.map((dto) => ({
-        recipeId,
-        name: dto.name,
-        amount: dto.amount,
-        unit: dto.unit,
-      })),
-    );
-    await this.ingredientRepo.save(ingredients);
-
-    return this.getIngredients(recipeId);
   }
 }
